// argefrontend/src/app/views/+basvuru/forms/model/doner-belgeler.ts
export class DonerBelgeler {
    ozGecmisBelgesi?: string;
    etikKurulOnayBelgesi?: string;
    anketOlcekFormu?: string;
    uzmanlikvbTeziBelge?: string;
    ekBelge?: string;

    BakanlikOnayi?: string;
    ProtokolOzeti?: string;
    GonulluOlur?: string;
    olguRapor?: string;
    butce?: string;
    calismayaKatilanMerkezler?: string;
    akisSemasi?: string;
    litaratur1?: string;
    litaratur2?: string;
    litaratur3?: string;


}